[SPARK-42282][PS][TESTS] Split `pyspark.pandas.tests.test_groupby`
### What changes were proposed in this pull request?
Move some slow tests from `test_groupby` to `test_groupby_slow`

### Why are the changes needed?
Sometimes `test_groupby` may take more than 20mins, this PR split it into two test files. So it maybe able to increase the parallelism.

in my local env:

before: (219sec)
```
(spark_dev) ➜  spark git:(ps_split_test_groupby) python/run-tests --testnames 'pyspark.pandas.tests.test_groupby'
Running PySpark tests. Output is in /Users/ruifeng.zheng/Dev/spark/python/unit-tests.log
Will test against the following Python executables: ['python3.9']
Will test the following Python tests: ['pyspark.pandas.tests.test_groupby']
python3.9 python_implementation is CPython
python3.9 version is: Python 3.9.15
Starting test(python3.9): pyspark.pandas.tests.test_groupby (temp output: /Users/ruifeng.zheng/Dev/spark/python/target/2cb4e2eb-e858-484f-8b7e-cf2463757c38/python3.9__pyspark.pandas.tests.test_groupby__tgagr3n9.log)
Finished test(python3.9): pyspark.pandas.tests.test_groupby (219s)
Tests passed in 219 seconds
```

after:  (116sec + 113sec)
```
(spark_dev) ➜  spark git:(ps_split_test_groupby) ✗ python/run-tests --testnames 'pyspark.pandas.tests.test_groupby_slow'
Running PySpark tests. Output is in /Users/ruifeng.zheng/Dev/spark/python/unit-tests.log
Will test against the following Python executables: ['python3.9']
Will test the following Python tests: ['pyspark.pandas.tests.test_groupby_slow']
python3.9 python_implementation is CPython
python3.9 version is: Python 3.9.15
Starting test(python3.9): pyspark.pandas.tests.test_groupby_slow (temp output: /Users/ruifeng.zheng/Dev/spark/python/target/84fb7cb6-074f-4ac6-9dd5-6aec929fb692/python3.9__pyspark.pandas.tests.test_groupby_slow__e_kgcdoh.log)
Finished test(python3.9): pyspark.pandas.tests.test_groupby_slow (116s)
Tests passed in 116 seconds
(spark_dev) ➜  spark git:(ps_split_test_groupby) ✗ python/run-tests --testnames 'pyspark.pandas.tests.test_groupby'
Running PySpark tests. Output is in /Users/ruifeng.zheng/Dev/spark/python/unit-tests.log
Will test against the following Python executables: ['python3.9']
Will test the following Python tests: ['pyspark.pandas.tests.test_groupby']
python3.9 python_implementation is CPython
python3.9 version is: Python 3.9.15
Starting test(python3.9): pyspark.pandas.tests.test_groupby (temp output: /Users/ruifeng.zheng/Dev/spark/python/target/e9aa7c59-6b1d-43c2-9e30-9fcb2b7172d9/python3.9__pyspark.pandas.tests.test_groupby__177eyt8w.log)
Finished test(python3.9): pyspark.pandas.tests.test_groupby (113s)
```

### Does this PR introduce _any_ user-facing change?
no, test only

### How was this patch tested?
updated tests

Closes #39849 from zhengruifeng/ps_split_test_groupby.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
